Add sentry rpc logging to wallet
