+++ apps/explorer/src/utils/api/DefaultRpcClient.ts
@@ -6,11 +6,9 @@ import {
     Connection,
     devnetConnection,
     localnetConnection,
-    JsonRpcClient,
-    type RpcParams,
 } from '@mysten/sui.js';
-import * as Sentry from '@sentry/react';
-import { type SpanStatusType } from '@sentry/tracing';
+
+import { SentryRPCClient } from './SentryRpcClient';
 
 export enum Network {
     LOCAL = 'LOCAL',
@@ -25,54 +23,6 @@ const CONNECTIONS: Record<Network, Connection> = {
         fullnode: 'https://fullnode-explorer.testnet.sui.io:443',
     }),
 };
-
-class SentryRPCClient extends JsonRpcClient {
-    #url: string;
-    constructor(url: string) {
-        super(url);
-        this.#url = url;
-    }
-
-    async #withRequest(
-        name: string,
-        data: Record<string, any>,
-        handler: () => Promise<any>
-    ) {
-        const transaction = Sentry.startTransaction({
-            name,
-            op: 'http.rpc-request',
-            data: data,
-            tags: {
-                url: this.#url,
-            },
-        });
-
-        try {
-            const res = await handler();
-            const status: SpanStatusType = 'ok';
-            transaction.setStatus(status);
-            return res;
-        } catch (e) {
-            const status: SpanStatusType = 'internal_error';
-            transaction.setStatus(status);
-            throw e;
-        } finally {
-            transaction.finish();
-        }
-    }
-
-    async request(method: string, args: any[]) {
-        return this.#withRequest(method, { args }, () =>
-            super.request(method, args)
-        );
-    }
-
-    async batchRequest(requests: RpcParams[]) {
-        return this.#withRequest('batch', { requests }, () =>
-            super.batchRequest(requests)
-        );
-    }
-}
 
 const defaultRpcMap: Map<Network | string, JsonRpcProvider> = new Map();
 /** @deprecated This shouldn't be directly used, and instead should be used through `useRpc()`. */
